+++ src/ui/views/ui_app.jade
@@ -159,7 +159,7 @@ block content
           err.style['text-align'] = 'center';
           err.style['background'] = '#F44336';
           err.style['padding'] = '25px';
-          err.innerText = '#{welcomeError}';
+          err.innerText = '#{i18n.translate('CORE-WELCOME_ERROR')}';
 
           document.body.innerHTML = err.outerHTML;
         }
